Recommendation Engine/requirements.txt
mysql-connector-python
pylint
pylint-django
pandas
